+++ app/renderer/components/index.js
@@ -6,9 +6,14 @@ import Input from 'components/Input';
 import Modal from 'components/Modal';
 import PageFooter from 'components/PageFooter';
 import PageHeading from 'components/PageHeading';
+import Scrollable from 'components/Scrollable';
 import Select from 'components/Select';
 import Spinner from 'components/Spinner';
+import Table from 'components/Table';
+import Tabs from 'components/Tabs';
+import TabPane from 'components/Tabs/TabPane';
 import Toggle from 'components/Toggle';
+import Tooltip from 'components/Tooltip';
 import TosContent from 'components/TosContent';
 
 export {
@@ -20,8 +25,13 @@ export {
   Modal,
   PageFooter,
   PageHeading,
+  Scrollable,
   Select,
   Spinner,
+  Table,
+  Tabs,
+  TabPane,
   Toggle,
+  Tooltip,
   TosContent,
 };