+++ styleguideHelpers/setup.js
@@ -2,6 +2,10 @@ import DemoCanvas from './DemoCanvas';
 import DemoSquare from './DemoSquare';
 import DemoFlex from './DemoFlex';
 
+import { TabPane } from '../app/renderer/components';
+
 global.DemoCanvas = DemoCanvas;
 global.DemoSquare = DemoSquare;
 global.DemoFlex = DemoFlex;
+
+global.TabPane = TabPane;